Only Pull From Enabbled Addresses
This is a problem pointed out to me by the_austria. The numbers would not
increment for disabled addresses, but were still included in the list being
pulled from, creating a discrepancy between the address listed for a number
and the address selected by that number.

The_austia's solution was to increment the numbers for disabled addresses.
I chose to create a new list that only includes enabled addresses, instead.
This way there are no gaps in the list.